# Use ISA-L to speed up gzip decompression