--- a/src/fastq.hpp
+++ b/src/fastq.hpp
@@ -25,7 +25,7 @@ public:
     void rewind();
 
 protected:
-    std::unique_ptr<AbstractIO> file;
+    std::unique_ptr<Reader> file;
     std::vector<std::vector<unsigned char>> saved_buffer;
     // if rewindable is false, the file cannot be rewound anymore and is consuming from saved_buffer (if it is not empty)
     bool rewindable;