--- a/src/iowrap.cpp
+++ b/src/iowrap.cpp
@@ -28,10 +28,14 @@ void RawIO::open(const std::string& filename)
     if(fd < 0) {
         throw InvalidFile("Could not open FASTQ file: " + filename);
     }
+
+    thread_reader = std::thread(&RawIO::preload, this, preload_size);
 }
 
 void RawIO::close()
 {
+    if(thread_reader.joinable()) thread_reader.join();
+
     if(fd != -1) {
         ::close(fd);
     }
@@ -40,8 +44,37 @@ void RawIO::close()
 
 int64_t RawIO::read(void* buffer, size_t length)
 {
-    ssize_t ret = ::read(fd, buffer, length);
-    return ret;
+    size_t actual_count = 0;
+    while(length > 0) {
+        size_t size_from_data = std::min(length, read_buffer.size() - read_buffer_copied);
+        memcpy(buffer, read_buffer.data() + read_buffer_copied, size_from_data);
+        buffer = (uint8_t*)buffer + size_from_data;
+        length -= size_from_data;
+        read_buffer_copied += size_from_data;
+        actual_count += size_from_data;
+
+        if(read_buffer_copied == read_buffer.size()) {
+            if(thread_reader.joinable()) {
+                thread_reader.join();
+                std::swap(read_buffer, read_buffer_work);
+                read_buffer_copied = 0;
+                if(read_buffer.size() == 0) {
+                    break;
+                }
+            }
+            thread_reader = std::thread(&RawIO::preload, this, preload_size);
+        }
+    }
+    return actual_count;
+}
+
+void RawIO::preload(size_t size)
+{
+    read_buffer_work.resize(preload_size);
+    auto actual_size = ::read(fd, read_buffer_work.data(), size);
+    if(actual_size != size) {
+        read_buffer_work.resize(actual_size);
+    }
 }
 
 void IsalIO::initialize()