--- a/src/iowrap.cpp
+++ b/src/iowrap.cpp
@@ -1,64 +1,60 @@
 #include "iowrap.hpp"
-#include "exceptions.hpp"
-#include <sys/stat.h>
-#include <sys/mman.h>
-#include <sys/types.h>
 #include <fcntl.h>
-#include <system_error>
+#include <sys/mman.h>
+#include <sys/stat.h>
+#include <sys/types.h>
 #include <cstring>
+#include <system_error>
+#include "exceptions.hpp"
 
-void GeneralIO::open(const std::string& filename)
-{
-    if(filename != "") {
+void GeneralIO::open(const std::string& filename) {
+    if (filename != "") {
         file = gzopen(filename.c_str(), "r");
-        if(file == nullptr) {
+        if (file == nullptr) {
             throw InvalidFile("Could not open FASTQ file: " + filename);
         }
     }
 }
 
-int64_t GeneralIO::read(void* buffer, size_t length)
-{
+int64_t GeneralIO::read(void* buffer, size_t length) {
     return gzread(file, buffer, length);
 }
 
-void RawIO::open(const std::string& filename)
-{
+void RawIO::open(const std::string& filename) {
     fd = ::open(filename.c_str(), 0);
-    if(fd < 0) {
+    if (fd < 0) {
         throw InvalidFile("Could not open FASTQ file: " + filename);
     }
 
     thread_reader = std::thread(&RawIO::preload, this, preload_size);
 }
 
-void RawIO::close()
-{
-    if(thread_reader.joinable()) thread_reader.join();
+void RawIO::close() {
+    if (thread_reader.joinable())
+        thread_reader.join();
 
-    if(fd != -1) {
+    if (fd != -1) {
         ::close(fd);
     }
     fd = -1;
 }
 
-int64_t RawIO::read(void* buffer, size_t length)
-{
+int64_t RawIO::read(void* buffer, size_t length) {
     size_t actual_count = 0;
-    while(length > 0) {
+    while (length > 0) {
         size_t size_from_data = std::min(length, read_buffer.size() - read_buffer_copied);
         memcpy(buffer, read_buffer.data() + read_buffer_copied, size_from_data);
-        buffer = (uint8_t*)buffer + size_from_data;
+        buffer = (uint8_t*) buffer + size_from_data;
         length -= size_from_data;
         read_buffer_copied += size_from_data;
         actual_count += size_from_data;
 
-        if(read_buffer_copied == read_buffer.size()) {
-            if(thread_reader.joinable()) {
+        if (read_buffer_copied == read_buffer.size()) {
+            if (thread_reader.joinable()) {
                 thread_reader.join();
                 std::swap(read_buffer, read_buffer_work);
                 read_buffer_copied = 0;
-                if(read_buffer.size() == 0) {
+                if (read_buffer.size() == 0) {
                     break;
                 }
             }
@@ -68,59 +64,60 @@ int64_t RawIO::read(void* buffer, size_t length)
     return actual_count;
 }
 
-void RawIO::preload(size_t size)
-{
+void RawIO::preload(size_t size) {
     read_buffer_work.resize(preload_size);
     auto actual_size = ::read(fd, read_buffer_work.data(), size);
-    if(actual_size != (decltype(actual_size))size) {
+    if (actual_size != (decltype(actual_size)) size) {
         read_buffer_work.resize(actual_size);
     }
 }
 
-void IsalIO::initialize()
-{
+void IsalIO::initialize() {
     isal_inflate_init(&state);
     state.crc_flag = ISAL_GZIP_NO_HDR_VER;
 
     isal_gzip_header_init(&gz_hdr);
 }
 
-void IsalIO::open(const std::string& filename)
-{
+void IsalIO::open(const std::string& filename) {
     fd = ::open(filename.c_str(), 0);
-    if(fd < 0) {
+    if (fd < 0) {
         throw InvalidFile("Could not open FASTQ file: " + filename);
     }
 
     struct stat _stat;
-    if(fstat(fd, &_stat) < 0) {
+    if (fstat(fd, &_stat) < 0) {
         throw std::system_error(errno, std::generic_category(), filename);
     }
     filesize = _stat.st_size;
 
     mmap_mem = mmap(NULL, filesize, PROT_READ, MAP_SHARED, fd, 0);
-    if(mmap_mem == MAP_FAILED) {
+    if (mmap_mem == MAP_FAILED) {
         mmap_mem = NULL;
-        if(errno == ENODEV) {
-            throw std::system_error(errno, std::generic_category(), "mmap is not supported on this file: " + filename);
+        if (errno == ENODEV) {
+            throw std::system_error(
+                errno, std::generic_category(), "mmap is not supported on this file: " + filename
+            );
         }
-        if(errno == ENOMEM) {
-            throw std::system_error(errno, std::generic_category(), "There not enough memory to open file: " + filename);
+        if (errno == ENOMEM) {
+            throw std::system_error(
+                errno, std::generic_category(), "There not enough memory to open file: " + filename
+            );
         } else {
             throw std::system_error(errno, std::generic_category(), "mmap failed to open file: " + filename);
         }
     }
-    mmap_size       = filesize;
+    mmap_size = filesize;
     compressed_data = reinterpret_cast<uint8_t*>(mmap_mem);
     compressed_size = mmap_size;
 
     // decompress gz header
-    state.next_in  = compressed_data;
+    state.next_in = compressed_data;
     state.avail_in = std::min(decompress_chunk_size, compressed_size);
-    auto pre       = state.avail_in;
+    auto pre = state.avail_in;
 
     int ret = isal_read_gzip_header(&state, &gz_hdr);
-    if(ret != ISAL_DECOMP_OK) {
+    if (ret != ISAL_DECOMP_OK) {
         throw std::runtime_error("Invalid gzip header found");
     }
     size_t processed_size = pre - state.avail_in;
@@ -130,72 +127,72 @@ void IsalIO::open(const std::string& filename)
     thread_reader = std::thread(&IsalIO::decompress, this, std::min(decompress_chunk_size, compressed_size));
 }
 
-void IsalIO::close()
-{
-    if(thread_reader.joinable()) thread_reader.join();
+void IsalIO::close() {
+    if (thread_reader.joinable())
+        thread_reader.join();
 
-    if(mmap_mem != nullptr) {
+    if (mmap_mem != nullptr) {
         munmap(mmap_mem, mmap_size);
     }
-    mmap_mem  = nullptr;
+    mmap_mem = nullptr;
     mmap_size = 0;
 
-    if(fd != -1)
+    if (fd != -1)
         ::close(fd);
     fd = -1;
 }
 
-int64_t IsalIO::read(void* buffer, size_t length)
-{
+int64_t IsalIO::read(void* buffer, size_t length) {
     size_t actual_count = 0;
-    while(length > 0) {
+    while (length > 0) {
         size_t size_from_data = std::min(length, uncompressed_data.size() - uncompressed_data_copied);
         memcpy(buffer, uncompressed_data.data() + uncompressed_data_copied, size_from_data);
-        buffer = (uint8_t*)buffer + size_from_data;
+        buffer = (uint8_t*) buffer + size_from_data;
         length -= size_from_data;
         uncompressed_data_copied += size_from_data;
         actual_count += size_from_data;
 
-        if(uncompressed_data_copied == uncompressed_data.size()) {
-            if(thread_reader.joinable()) {
+        if (uncompressed_data_copied == uncompressed_data.size()) {
+            if (thread_reader.joinable()) {
                 thread_reader.join();
                 std::swap(uncompressed_data, uncompressed_data_work);
                 uncompressed_data_copied = 0;
-                if(uncompressed_data.size() == 0) {
+                if (uncompressed_data.size() == 0) {
                     break;
                 }
             }
-            thread_reader = std::thread(&IsalIO::decompress, this, std::min(decompress_chunk_size, compressed_size));
+            thread_reader =
+                std::thread(&IsalIO::decompress, this, std::min(decompress_chunk_size, compressed_size));
         }
     }
     return actual_count;
 }
 
-void IsalIO::decompress(size_t count)
-{
+void IsalIO::decompress(size_t count) {
     uncompressed_data_work.resize(std::max(count, previous_member_size));
     uint8_t* ptr = uncompressed_data_work.data();
 
-    while((uintptr_t)(ptr - uncompressed_data_work.data()) < (uintptr_t)count) {
-        if(compressed_size == 0) break;
+    while ((uintptr_t) (ptr - uncompressed_data_work.data()) < (uintptr_t) count) {
+        if (compressed_size == 0)
+            break;
 
-        size_t actual_input_size     = std::min(decompress_chunk_size, compressed_size);
+        size_t actual_input_size = std::min(decompress_chunk_size, compressed_size);
         size_t output_size_available = uncompressed_data_work.data() + uncompressed_data_work.size() - ptr;
 
-        state.next_in  = compressed_data;
+        state.next_in = compressed_data;
         state.avail_in = actual_input_size;
 
-        state.next_out  = ptr;
+        state.next_out = ptr;
         state.avail_out = output_size_available;
 
         auto ret = isal_inflate(&state);
-        if(ret != ISAL_DECOMP_OK) {
+        if (ret != ISAL_DECOMP_OK) {
             throw std::runtime_error("Error encountered while decompressing");
             exit(-1);
         }
 
         size_t processed = actual_input_size - state.avail_in;
-        size_t decomp    = (size_t)(uintptr_t)(state.next_out - ptr);
+        size_t decomp = (size_t) (uintptr_t) (state.next_out - ptr);
 
         compressed_data += processed;
         compressed_size -= processed;