@telmin 2023-04-27
I am trying to fix CI, but I can't figure out why it SEGVs on MacOS and why the build fails with pythonbindings. Can you please give me some advice?
I tried to create a similar environment locally for pythonbindings and the build passes. it may be a problem with the SOURCE_DIR variable, but I am not sure what the cause is...