@marcelm 2023-04-27
> I tried to create a similar environment locally for pythonbindings and the build passes. it may be a problem with the SOURCE_DIR variable, but I am not sure what the cause is...

For testing the Python bindings, you should compile via `pip install`. So something like this:
```
python3 -m venv .venv
.venv/bin/pip install .
```
I was able to reproduce the error when I did the above. But as I said, I will try to help with figuring out that issue.