@marcelm 2023-04-27
>     1. Usage of ISA-L needs to be optional since it is not available everywhere (Apple Silicon for example)

Hm, it seems to be available for aarch64, so I take that back.