@marcelm 2023-04-27
Once again regarding this:

> 2\. We can tackle this later, but should keep it in mind: We need to make it possible to dynamically link to `libisal` so that an already installed one can be used. So when building the Conda package, for example, we would just add `isa-l` as a dependency.

Maybe it is actually better to always assume that libisal is available on the system. Looking at this [isal-l wiki page](https://github.com/intel/isa-l/wiki/Ports--Repos), it’s clear that libisal is available on many Linux distributions and on macOS via brew. So this is probably simpler and more reliable than using `ExternalProject`.