@marcelm 2023-04-27
It got this working using the system isa-l by replacing `ExternalProject_*` with this:
```
find_package(PkgConfig REQUIRED)
pkg_check_modules(ISAL REQUIRED IMPORTED_TARGET GLOBAL libisal>=2.30.0)
target_include_directories(salib PUBLIC src/ ext/ ${PROJECT_BINARY_DIR} ${ISAL_INCLUDE_DIRS})
target_link_libraries(salib PUBLIC ZLIB::ZLIB Threads::Threads zstr::zstr PkgConfig::ISAL)
```
And in `iowrap.hpp`, I had to change `#include "igzip_lib.h"` to `#include "isa-l/igzip_lib.h"`. I then installed libisal with `sudo apt install libisal-dev` and strobealign compiled fine, and even `pip install .` worked.