@telmin 2023-04-28
Thanks for the review. I think I was able to fix all the points you made.
I should have used isa-l from the package manager, as you pointed out.

I am not sure what is causing it to SEGV under MacOS...