@marcelm 2023-05-03
Regarding the macOS failure, this is the error message:
```
Total time writing index: 0.02 s
tests/run.sh: line 56:  2953 Segmentation fault: 11  strobealign -r 150 -i tests/phix.fasta
```
The `strobealign -i` command does not even read from any input FASTQ files, and it crashes right at the end shortly before it would exit anyway. Maybe some cleanup that doesn’t work as expected?