@telmin 2023-05-11
The reason I am double-buffering the gzip decompression is to speed up the process; trying to do gzip decompression without double-buffering would be very slow, so I am buffering to avoid this.

But you are right, I think we should properly consider if we really need GB of Buffer.
So I took some benchmarks.
In the following benchmark, decompression is done n bytes at a time and the decompressed buffer is allocated to 4*n bytes.

* 256KB / 1MB
```
Total time reading read-file(s): 211.97 s.
```

* 1MB / 4MB
```
Total time reading read-file(s): 109.14 s.
```

* 2MB / 8MB
```
Total time reading read-file(s): 82.67 s.
```

* 4MB / 16MB
```
Total time reading read-file(s): 85.07 s.
```

* 8MB / 32MB
```
Total time reading read-file(s): 88.06 s.
```

* 16MB / 64MB
```
Total time reading read-file(s): 85.09 s.
```

* 32MB / 128MB
```
Total time reading read-file(s): 90.15 s.
```

* 64MB / 256MB
```
Total time reading read-file(s): 91.16 s.
```

* 128MB / 512MB
```
Total time reading read-file(s): 93.04 s.
```

* 256MB / 1GB
```
Total time reading read-file(s): 92.55 s.
```

On my machine, with the proper buffer size, it seems to generally finish in around 80sec. 256KB unit decompression does not seem to cover up the decompression time. So I did the decompression in 2MB increments and changed the buffer size to 8MB for decompressed data.