@marcelm 2023-05-26
Hi, I’m back from traveling, sorry for the delay. Thanks for the update, and great that you could even fix the macOS issue.

I’m trying to understand `UncompressedReader` and `IsalGzipReader`, but am a bit slow. I think a couple of comments would have helped. But can you help me for now by explaining a couple of things:
- What are `read_buffer` and `read_buffer_work` and what is the difference?
- Why does `UncompressReader` use a separate thread?
- Is it correct that a thread is created and joined for each read chunk? (I guess it’s ok because thread creation is efficient.)

I also wonder whether using `mmap()` for the input is appropriate. I’d tend to use that for files that need random access, but the input FASTQ needs to be streamed only once, so I don’t see why it would be necessary to map it in its entirety into memory. It also appears to make the code more complex as you now need to find out what the file size is in advance.