@marcelm 2024-04-22
This has been sitting here for a while, but it would still be nice to have in strobealign. I made the modifications I think are necessary to merge this and opened #418 instead.

@telmin I squashed your contribution to a single commit (preserving authorship information).

I’m closing this now in favor of #418.